JDO-817 removing compiler warnings: redundant type cast

+++ tck/src/main/java/org/apache/jdo/tck/models/fieldtypes/TestFieldsOfByte.java
@@ -76,8 +76,8 @@ public class TestFieldsOfByte extends JDO_Test {
     {
         Transaction tx = pm.currentTransaction();
         int i, n;
-        Byte firstValue = Byte.valueOf((byte)Byte.MIN_VALUE);
-        Byte secondValue = Byte.valueOf((byte)Byte.MAX_VALUE);
+        Byte firstValue = Byte.valueOf(Byte.MIN_VALUE);
+        Byte secondValue = Byte.valueOf(Byte.MAX_VALUE);
         tx.begin();
         FieldsOfByte pi = new FieldsOfByte();
         pi.identifier = 1;
