+++ tck/src/main/java/org/apache/jdo/tck/pc/fieldtypes/HashtableStringKeyCollections.java
@@ -749,7 +749,7 @@ public class HashtableStringKeyCollections {
 
         public String toString() { return this.getClass().getName() + ": "  + identifier;}
 
-        public int hashCode() { return (int)identifier ; }
+        public int hashCode() { return identifier; }
 
         public boolean equals(Object other) {
             if (other != null && (other instanceof Oid)) {
