+++ tck/src/main/java/org/apache/jdo/tck/pc/fieldtypes/TreeMapStringValueCollections.java
@@ -699,7 +699,7 @@ public class TreeMapStringValueCollections {
 
         public String toString() { return this.getClass().getName() + ": "  + identifier;}
 
-        public int hashCode() { return (int)identifier ; }
+        public int hashCode() { return identifier; }
 
         public boolean equals(Object other) {
             if (other != null && (other instanceof Oid)) {
